pyproject.toml
[tool.poetry]
name = "logscalescim"
version = "0.1.0"
description = ""
authors = ["Ryan Faircloth <ryan@dss-i.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"
gql = {version = "^3.4.0", extras = ["requests"]}
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.2"
opentelemetry-instrumentation-flask = "^0.49b2"
opentelemetry-instrumentation-requests = "^0.49b2"
opentelemetry-instrumentation-logging = "^0.49b2"
opentelemetry-exporter-otlp = "^1.28.2"
opentelemetry-api = "^1.28.2"
opentelemetry-sdk = "^1.28.2"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
